libstagefright: Release output buffer when file limit is reached in recording
There is a memory leak in video recording when the
file size or duration limits are reached. The output
buffer copy is not released before exiting the track
thread in the mpeg4 writer.

To avoid the memory leak, the buffer copy is released
when a limit condition is detected in the mpeg4 writer.

Change-Id: I7997bcab84836434038f9c9de0901307c0f01054
CRs-Fixed: 315255