libstagefright: Use writer thread for video only recording.
The MPEG4 writer is currently writing the frames to the file in
the video track thread when doing video only recording. This causes
frame drops when the call to write to the file takes too long.

With this change, frames are written to the file from the writer
thread on video only recording, preventing the video track thread
to get blocked on writes.

Change-Id: Iad38532e6b2ca5449f51af5869be1049ea1f0824
CRs-Fixed: 341670