frameworks/av: Fix for auto resume issue when QCELP clips are paused
-Issue: The issue happens with a timing where the playback is paused
and there is a wait on the buffer filled. The wait timed out as the
playback is in pause state notifying TIMED OUT error and as a result
the playback is abruptly closed follwed by the next song playback
from the list.

-Test Scenario: Device switch from headet to speaker when playing
QCP clips and wait for 5 seconds. The issue is observed at random

-Fix: Handled the TIMEDOUT by ignoring the timed out error if the
playback is in pause state.

(cherry picked from commit a92be8395f0f5eaafbd9fdd4d9c9a470c96612e3)

Change-Id: I4e46e575468e79ec767e8eebc3f2d380ae7fe188