libstagefright: Exit extractor thread before pause is issued to audio sink
On MPQAudioPlayer reset, pause is called for audio sink, which pasues the
Audio dirver, now if the there is no buffer available in the empty queue
Then the extractor thread in MPQAudioPlayer is blocked on write call.
Exit the extractor thread before the pasue is issued to audio sink.

CRs-Fixed: 385015
Change-Id: I81d58113d12aad1876f92aef31041552a5c95d26