libstagefright: Fix EOS - Seek issue.
Seeking to EOS in tunnel mode is causing no audio. This is because
zero length buffer is not handled properly.
Fix is to not send zer length buffer to driver if the queue is
already empty.

Conflicts:

	media/libstagefright/TunnelPlayer.cpp

Change-Id: I0c3977c6e8638aada17645cb50dba0bebc9d5509
(cherry picked from commit 6dcf8f86f760b3633fcd9305336e03a11f3e8747)