libstagefright: Remove the 0 byte write to driver at EOS
Since the break in the compressed driver is tracked by the
app and hw pointer so the 0 byte write hack is no more needed.
This change caused an issue when writing of 0 bytes is taking time
resulting in the miss of the signal to break out of the lock.

CRs-Fixed: 405837
(cherry picked from commit f54ca9bb0ffd17db2f227997e5ff0a128f9f629d)

Change-Id: I5fe69f10133f6eac5acbed22eb0590773d28fef5

Signed-off-by: Rajesh Puvvala <puvvar@codeaurora.org>
Signed-off-by: Srikanth Uyyala <suyyala@codeaurora.org>