audioflinger: Fix the routing of the devices during pause resume
For the WFD, when BT is disconnected audio stream goes in the pause
and when it is resumed it causes a crash.

When the stream is resumed the resume to stream is sent followed by the
startOutput which performs the routing. This causes the stream to be routed to BT
and then perform a switchDevice. This causes the failure as proxy read device has
already been opened by the WFD.

First start the audio output followed by the stream resume, this will ensure the stream
is resumed with the updated device.

(cherry picked from commit 317fcb8b9d95f810721d35e101fc944313adeeab)

Change-Id: Iaad45a84632d1fd8e326db4f02d091d3c92c2052