audioflinger: Fix for improper LPA playback after seek.
-LPA playback does not happen in continuous manner when seek is
performed rapidly and effects is turned on/off.
-After seek, mSkipwrite flag is set to skip the current buffer either
in bufferavailable or write of HAL. If returned from write there is a
mismatch in buffer maintenance between flinger and hal as flinger
is not aware of buffer skip and has already pushed back the buffer
into mEffectsPool. Also effectsPool operations in flush() are not
protected by any lock.
-Handle error condition in skip case and re-arrange buffers to
original state in mEffectsPool to avoid any mismatch.

Change-Id: I4d9bc979bc10cb9fa131575634f80ac117a6e635
CRs-Fixed: 463272