frameworks/av:Don't wait for filled buffer if there IS input data and time-out error
Issue: If OMX client is waiting for filled buffers from codec
after sending pause to it,it may lead to deadlock, as codec goes
to 'SLEEP' once it changes state to pause and after that process
a fill buffer request. So client keeps waiting for filled buffers.

Fix: Do not wait for filled buffers if we are in paused state and
we get a time-out although there is a input data to decoder.
Change the waitForBufferFilled_l to report a error if we get a time-out.

Change-Id: I06480fe19a19de441fc29248ac3ce9acd3dfda99
CRs-Fixed: 460520