libstagefright: Fix the AVsync issue for the .ts clips
For the .ts clips we also receive the audio pts from the clip.
Currently the pts values are not taken into account which results
in the loss of AVsync for a AV clip

Fix by tracking the pts values in the mediaTimeMapping which is used
to evaluate the lateness between the audio and video packets.

Change-Id: I67961f300fd2b0d5ba38da596782fdd56c63ad01