audioflinger: fix volume not updated in pause state of direct track
When in pause state, the state of the track is moved to inactive.
The volume update is discarded. Hence, cache the volume and send
it when the state is moved to active state.
Make the stream to NULL once stoped

CRs-Fixed: 490592
(cherry picked from commit a6c131f6c684359e7ab332c4d0f83fd20e8163d4)

Change-Id: I3a09efba72449a737916311d913f9864f4b2fb03