libstagefright: update crop and stride synchronously on resolution change
When crop change is detected, save the target crop/stride and apply
them just before the buffer is (potentially) rendered.
Applying the updates right away may cause artifacts as buffers with
previous resolution are not rendered yet.

This approach assumes that there will not be a crop change within
the detection and application of current crop.

CRs-fixed: 554120

Change-Id: I8ab8b18482862afcccbb1f72bb13768010a8ce47