Audio: Use temporary flags to force fast flag
-Issue: Camcorder start tone gets encoded in video encoded clip.
-Rootcause: processAudioBuffer() in AudioTrack waits for priority
boost for almost 630ms after which it starts reading data from
AudioPlayer.
Priority Boost is not received because AudioTrack forces fast flag
for this stream. Due to this, there is no request for data from
AudioPlayer to OMXCodec leading to stall in vorbis decoder. This
results in recording of the start tones even after the 300ms
initial delay.
-Fix: Since the idea is to use only fast output profile, we can
avoid manipulating flags completely and use new flags in
case of getOutput. This will avoid changing original flags.

CRs-Fixed: 555291
Change-Id: I34fdbb3be03f5a3ce7dc70438b722e25f951f561