libstagefright: Set initial buffer size for H264 codec
For Google h264 decoder, the input port buffer size is
set to 8192 bytes, and for usecases where data source
does not specify input buffer size, there is a crash on
assert when the buffer size is larger than 8192 bytes.

Added change to set input buffer size for H264 to 80k.

CRs-Fixed: 619869

Change-Id: I5bc177cc5078b2acfc930ee2746960dc26ab9662