frameworks/av: Use the software decoder for the Session from AudioCache
Some of the apps uses the SoundPool for the compressed format like
MP3. Current implementation always assumes the compressed format to go via
the tunnel mode.
So the MP3 audio is sent as PCM data causing the noise.

Since the SoundPool requires the decoding in the NT mode,
hence intilaize the software decoder irrespective of the formats.

CRs-Fixed: 620275
Change-Id: I970022f8b14214aa351093839766f481d5268697