audio policy: Modify check to take care of updated timestamp
-There is a delay to play the key tone after hanging up a
voice call when tried immediately.
-StartOutput called during start of tone is getting blocked for 2sec.
AudioCommands queue in AudioPolicyService is getting corrupted due
to which fm_volume command with a delay of 2sec supposed to be at
the end of the queue moves to the top.
when timestamp of current command is compared to the other audio
commands in the queue, it uses the original timestamp of the
command through out and iterates through the queue accordingly.
At the end of the insertCommand_l function timestamp of current
command can be updated if there is an already existing timestamp
of the same command with a greater value, in this case fm_volume.
Due to this, the insertion of the command need not be in the
increasing order of timestamps of the audio command queue disturbing
the order of the queue.
-Modify comparison condition in insertCommand_l to take care of
this to avoid such issues.

Change-Id: Idd3c8e7007b4eac817398efcf59e06ee50a5e6ba
CRs-Fixed: 605091