stagefright: yield callbackDispatcher to writer only for video
Letting Writer run increases the probability of dispatcher
not getting the lock when it comes back with the next message.
This adds an extra sleep-wakeup and increase power consumption.

Since Audio dispatcher handles more messages, this behavior
is profound in case of audio. Contain yielding business to
video only since this optimization is not required for audio.

CRs-Fixed: 651233
Change-Id: Ie24be4522b29fa191a88a777293d39f8e7c91353