stagefright: increase buffers with nativewindow
to avoid frame-drops due to latencies in rendering.

If (minUnDequeued) buffers are caught in display pipeline,
dequeueBuffer will block, which delays subsequent queuebuffer,
and this repeats until a buffer is canceled (frame drop).
This issue is more prominent with high resolution content
on secondary displays.
Adding additional buffers increases the stagger in display
and makes finding a free buffer more likely.

CRs-Fixed: 660162
Change-Id: I047525025cb683e1a6a678b907b50dde208fdcc9