RTSP: Correct the data type in the check
The check for a race condition in RTSP flush has a mismatch
in its data type.
UNKNOWN_ERROR is used as a positive value and isn't equal to
the return value of feedMoreTSData().
Typecast it to fix this mismatch.

(cherry picked from commit 5818610e0be14d9c5d750463dee85d29d422eaf7)
Change-Id: If4f3e26cd926581e63fd29d085fb9c6f1c7fa9cf