stagefright: httplive: Fix switching for HLS live use-cases
Playlists can periodically update during HLS live streaming, and
the first sequence number can change accordingly. When switching,
this leads to a mismatch in first sequence numbers that causes
selection of the wrong segment, as the passed segmentStartTimeUs
value is relative to the old first sequence number.
For a/v fetchers, pass the previous a/v fetcher's sequence number to
each new fetcher to enforce continuity of a/v streams.

CRs-Fixed: 771954
Change-Id: I26d8118dbcb7300571a5b3fb5d0808c5a0c08aa6