httplive: Create a copy of last enqueued metadata
Create duplicate of metadata in queueAccessUnit so that it is
available even after all the buffers are erased from packet source.

During a bandwidth switch httplive streaming Source uses last
enqueued timestamp as the start time for the new streams, and this
switch can occur at a time when all the packets are dequeued from
the current packet source. This is one of the scenarios when the
last enqueued time was showing a invalid timestamp. Creating a
copy will retain the timestamp value until the packet source is
active.

CRs-Fixed: 808563
Change-Id: I4d4ee700705cee58773da4660f8769f56018f9e4