httplive: Do not delay fetcher when resuming
During a switch resumeUntil is called on old fetcher so
that it continues to download until the timestamp gap between
old and new fetchers is filled. If this is posted with a delay
new fetcher is scheduled first and it continues to download
second segment increasing the switch latency.To avoid this
resume fetcher download immediately.

CRs-Fixed: 771954
Change-Id: Ibdde95e016d83286dceed3f001b9a1d8b56b0367