httplive: Enable rapid switch down on bandwidth drop
Since bandwidth estimation relies on an average, it is not
immediately responsive to bandwidth drops. This can lead to
selection of a higher bandwidth index than can actually be
supported. Once the bandwidth estimate stabilizes, the new
fetcher is stuck buffering and is guarded by the mSwitchInProgress
flag often leading to a visible freeze.

If the bandwidth index drops while in the middle of a switch and
there is not sufficient data buffered to handle the switch latency,
cancel the previous switch and switch down to the correct index.

CRs-Fixed: 771954
Change-Id: I434d2f2a616f2542f193c4272ef3cc7c77b3eb35