audio: Fix memory leak in mediaserver
startOutput is forcing device switch even if other opened output
is not active.This is resulting in creation of new AudioPatch.
Where as stopOutput doesn't force device switch , if other opened
output is not active.This is causing memory leak because AudioPatch
created during startOutput is not getting deleted as part of
stopOutput. Fix is to force device switch from startOutput, if
other output is active.

Fixed another memory leak issue in IAudioPolicyService

Change-Id: I7c4769f3fa74a0fa754a15987864dc07e16afa00