stagefright: httplive: Estimate file size for chunked files
Clips using HTTP chunked transfer encoding do not return a valid
content-length in the http response. This makes it difficult to
allocate properly-sized buffers for chunked content, and many
copies may be necessary as the buffer size is dynamically increased.

Use the bandwidth & target-duration metadata present in variant
streams to estimate the file size of a segment with a specific
bitrate to avoid unnecessary copies.

CRs-Fixed: 771954
Change-Id: Ia1022883a3dcfccceca99e4129352696eaa273c4