audiopolicy: Do not route VoIP call to HDMI
getDeviceForStrategy is returning AUDIO_DEVICE_OUT_AUX_DIGITAL
even  when setForceUse is called with FORCE_NONE(earpiece)
during VOIP call. Actual Intention is to route audio for phone
strategy to AUX device even after setForceUse is called with
FORCE_NONE when not in voice call. It is supposed to exclude
VOIP call too

Added isInCall check so that it returns EARPIECE when
setForceUse is called with FORCE_NONE

CRs-Fixed: 793649

Change-Id: I88d515c351f066305f9eed240b1fe5f60ef34f85