libstagefright: Support PCM channel mask update
For PCM offloaded content that has a different
channel count when compared to what is
specified in the header, playback fails.
Audio HAL expects the updated channel mask
to be sent to it. Fix this by overwriting
the channel mask retrieved from the header
of the file format with the channel mask
retrieved based on the decoders output

Change-Id: Idf6f1efeeb764660bbb91a63e9f351d64d35a791
CRs-Fixed: 825717