audiopolicy: fix memory corruption in libaudiopolicymanagerdefault
Fix memory corruption in libaudiopolicymanagerdefault.
Definition of class AudioPolicyManager in libaudiopolicymanagerdefault
is not matching with the definition used in libaudiopolicymanager,
though the header file used by both the libs are same, compilation flags
used for them are different leading to different definitions.
Due to this difference in definition binding object from one module to
another results in memory corruption.
Use correct CFLAGS for both the libs to avoid this discrepancy and hence
prevent memory corruption.

Change-Id: I19e6f457cecae75a544eafae00b7397829b083ba