audiopolicy: fix to play 6.1 channel ALAC clips
For 6.1 channel ALAC clips, the compress offload profile is
not found even though the channel mask is defined in the
audio_policy.conf file. This is because the channel mask enum
is defined under an improper feature flag which is not enabled
and hence the profile's channel masks are not getting
enumerated properly.

Remove featurization of the channel masks. Also remove the duplicate
entry for QUAD. This change is needed for 6.1 channel ALAC clips playback.

CRs-fixed: 870544
Change-Id: I6b820776c0dc6e68a402886f0931439edab24a8b