audio: edit declaration of APE compression level
APE compression level is a 16 bit integer but is currently defined
as a 32 bit integer. This causes incorrect parsing of the
compression level from the structure populated by the parser and
some legitimate clips fail to play as they are rejected.

Declare compression level as a 16 bit integer to fix the issue.

Also, unsupported APE clips are now rejected in the OMX component,
so remove the handling in framework.

CRs-Fixed: 874425
Change-Id: I2b301cee6977d4e40bf1a1055e2aeb787a44b842