NuPlayer: Defer seek during offload teardown
On an offload teardown, the renderer is flushed and the decoder is
flushed+shutdown. A seek is issued on the parser to resume
playback from the last known playback position.

The flush on the renderer triggers an input buffer request in the
decoder module. This request ends up filling up the aggregate buffer
but all of it is lost when the pending flush+shutdown command is
handled by the decoder.

When playback is resumed, the new decoder instance gets data at a
much farther position than expected leading to an unexpected
playback position.

To fix this, defer seek to after the new decoder instance is created.

CRs-Fixed: 874120
Change-Id: I3c24cc91f29297407a9a2589b60a3ebeb8f5f64b