stagefright: httplive: Optimizations in Bandwidth estimation
The Bandwidth estimation routine needs to detect rapid downswitch
in the bandwidth.

Below changes are done in addBandwidthMeasurement()

- Keep only last 5 seconds worth of data samples for the estimation.
- Erase old samples if the total transfer time is greater than 5sec
  and total number of samples greater than 8.

Below change is done in estimateBandwidth()

- keep the latest 3 items when the total time is greater than 5s

Added setCustomBwEstimate() API in HTTPBase so that it gets called
from LiveSession and the changes will be applicable only for HLS and
when persist.sys.media.hls-custom is set to true.

Change-Id: Ia0c511143b61d618796d8f9ad6e8ce25197f6d5a