MPEG4Extractor.cpp: Add check for size == SIZE_MAX.
If size == SIZE_MAX, the line:

  uint8_t *buffer = new (std::nothrow) uint8_t[size + 1];

ends up allocating zero bytes, which is obviously incorrect.

This is conceptually a cherrypick of commit
b2d33aee5122c91a59c2a676c0b89ad340232450 , but specifically for
Android 4.1 through Android 4.4. In Android 5.0, new code
was introduced which caused the function parseMetaData()
to be renamed.

Bug: 23031033
Change-Id: Ib34e740f3292a484f8a24e513c1cce58f2f33ecb