audio: Fix VoIP direct output profile selection
Ringtone mute is observed in LINE app MT call over BT headset.
VoIP direct output isn't returned even though the required flags
are set. In this scenario, DIRECT flags are set but track is
getting primary output. This is resulting in incorrect values
when renderer queries for getPosition(). Fix this to return
VoIP direct output when the appropriate flags are set.

CRs-Fixed: 893558
Change-Id: I4fe8bc05455e337c24452e49be20ec36eec9c85d