audio: clear vorbis anchor buffer everytime aggregation finishes
vorbis anchor buffer is not reset to NULL when one round of aggregation
finishes, so it's still available in the next round. This leads to an
assertion failure, because mAnchorBuffer should only appears as a first
small buffer.

Fix is to ensure vorbis anchor and header buffer get recycled upon
aggregation finishes.

Change-Id: If07bbf97bd035eadd756703498bf1d6ff68a34c4
CRs-Fixed: 909091