audio: Do not stop SoundChannel for NEW_IAUDIOTRACK event
SoundChannel re-uses AudioTrack if the channel is initialized
previously for the same sampleId. This is causing issues
when the underlying AudioTrack is invalidated by Audio Policy.
When start() is called on an invalidated AudioTrack, track
is restored and it triggers an event to SoundChannel to stop
the track

Fix is not to stop SoundChannel in this case as track
is already restored and started succesfully for deep
buffer path

Change-Id: I891e447fb0760fa5c93319769b1e4d0463cebbb7