Revert "audiopolicy: optimize voice call start latency"
This reverts commit 040d7ecbd663420e39af7f0f56437d381ca310ac. Pops
heard when MO call is initiated while music playback is in progress.
This is due to a new device is set up for the output even before the
leftover buffers were drained out on the previous device.
This change is to make sure to add proper delay while device
routing, which is equal to twice of Output descriptor latency.

CRs-Fixed: 906398
Change-Id: I08470611172b291d2dea413ed822226777f5d157