httplive: fix corner case during buffering
If the duration in packetSource is equal to the target
duration, LiveSession will still keep the buffering status
not to feed decoders the buffers. Meanwhile PlaylistFetcher
stops to fetch more data because the duration is not lower
than the target duration. Then the playback is freeze.

Add the equal sign to fix this corner case.

Change-Id: I22c288ad99415852514ec41f9a61451813b78f70
CRs-Fixed: 920830