audiopolicy: check only current phone state in isInCall
Sometimes isInCall is returning true even when phone state
is not MODE_IN_CALL. Issue is when BT is connected and receive
MT call. Once the call is connected DTMF tone is not heard
on BT. When setPhoneState is called with RINGING, getDeviceForStrategy
will decide the device based on isInCall state. But isInCall is
returning true even when phone state is RINGING.

Fix is to check only the current phone state in isInCall
CRs-Fixed: 932873

Change-Id: I06d00c4be44f8134f1ba66fb964f5b1aff0faba2