libstagefright: Stop Track before exceeding max duation in MPEG4Writer
When max file duration is specified by user, the duration of recorded clip
might exceed this by couple of milliseconds. This is because we check on
the Track's duration too late.
So check Track's estimated duration before writing current sample,
if it exceeds max duration, drop this sample and stop recording.

Change-Id: Ie1bf83bb22c852c46e8e53a24109962c057965b6
CRs-Fixed: 964379