Stagefright: Revert initiateShutdown timeout from 30s to 3s
This timout value was changed from 3s to 30s due to some software
codecs operations took a long time, if OMX node was force-released
after timeout, there would be a race condition. But this issue has been
fixed now, and the timeout value 30s is too long, it will cause Camera
ANR in some condition, so revert it.

Change-Id: Ib0c27f0f7095df327c12572b7d0ea80481d2beb2
CRs-Fixed: 1036171