StageFrightRecorder: fix calculating paused duration
If pause is issued before start time offset is elapsed, the
paused duration will include the remaining offset time. This
will make kKeyTime that retrieved from AudioSource to be less
than the paused duration. This situation causes an assertion.

Set the paused time to the time when actual recording started
if pause is issued before offset time is elapsed, this can
ensure the calculated total paused duration is less than the
kKeyTime.

Change-Id: I49b3409c530c4d728fb5ff8bf9124909d03ffb1a
CRs-Fixed: 1062825