audioflinger: pause direct output when buffer timeout
Previously, direct output can be paused when underrun is observed
for the first time, this can interrupt the playback even there is
PCM data cached in the driver, and underrun may be recovered
within on the next loop cycle. Pausing the HAL is unnecessary at
this time.
Only do HAL pausing when track is going to be removed for buffer
timeout, allow the client to fill the buffer for the next loop
cycle.

Change-Id: I3707d883d0daa4ffd2baa9bd69aebf4bf9ddef4e
CRs-Fixed: 1094472