libstagefright: Add forced shutdown option in ACodec
Add a forced shutdown option in ACodec which will do proper
state transitions and memory cleanups. This will be required
if ACodec got stuck in IdleToExecuting state and never received
an expected response from the OMX component. In that case, the
shutdown call will be infinitely stuck. This change posts a
forced shutdown message after a timeout which forcefully changes
the state back to Loaded.

Change-Id: I48e00e463e6114238c801a767574f9cc6521c097