media : Changing time stamp manipulation in HFR recording.
Earlier during pause resume operations the pause duration
was multiplied with the ratio of capture FPS and the desired
frame rate of the recorded video. This logic led to situations
where if the pause duration was less than the above quantity,
on subtracting, the value would come negative. This led to
elongation of the length of the video beyond the desired length.
Also the earlier logic led to miscalculation of the timeoffset
that needs to be subtracted from the timestamp from camera. This
led to elongation of time length as the value subtracted was
much lesser than what was required.

Change-Id: I98cdb14bb2b9c86013df9b2c8f2e558f184b633e