ACodec: Add generation logic for forced shutdown
Spurious handling of forced shutdown can happen when codec is
not released after stop and is in continuous start-stop. Add
a generation logic to avoid this. Attach the generation variable
every time while sending the forced shutdown message, and increment
it if shutdown is properly handled before timeout. Compare the
generation variable with current value before handling the forced
shutdown message, and handle only if it matches.

Change-Id: I15c799a798837be683abd619c1790ac3bef16001