Avoid presentationComplete loop when hardware is paused..
When client calls pause->stop->start and immediately
calls stop on AudioTrack, prepareTracks_l of
directOutputThread loops indefinitely with
presentationComplete call. Reason being framesWritten
does not get incremented in this case and
framesWritten never reaches presentationCompleteFrames.

Fix is to avoid presentationComplete loop indefinitely
when hardware is paused and reset the track and
flush the hardware.

CRs-Fixed: 2196647
Change-Id: Ie398a5f39c1784529709426a4a9c036f78687b57