audio: calculate the max buffer size runtime
Currently kMaxBufferSize is predefined as 30720,
but this is valid only for camcorder use-case and
not for other use-case, but as this is predefined
calculations for other than camera use-case go
wrong, such as frameCount and record latency, which
leads to issues such as data drop at the begining
due to high record latency.
So instead predefining the MaxBufferSize calculate
it only for camcorder use-case and use standard AOSP
value otherwise.

Change-Id: Ie7d43ddfdcb2b13a8a5eaa12cc8a233ac0d9470d